app/requirements.txt
numpy<2.0
chromadb==0.5.0
sentence-transformers==3.0.0
pypdf==4.2.0
python-docx==1.1.2
python-magic==0.4.27
markdown==3.6
watchdog==4.0.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3